Move onError callback and <ErrorToastContent /> component from @stlite/sharing to @stlite/common-react
