+++ packages/common-react/src/toastify-components/index.ts
@@ -1,2 +1,3 @@
 export * from "./callbacks"
 export { default as ToastContainer } from "./ToastContainer"
+export { default as ErrorToastContent } from "./ErrorToastContent"
