--- packages/sharing/src/components/ErrorToastContent.module.css
@@ -1,4 +1,0 @@
-.codeContainer {
-  overflow: scroll;
-  max-height: 300px;
-}
